+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,52 @@
+# [@adobe/spacecat-shared-content-client-v1.3.39](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.38...@adobe/spacecat-shared-content-client-v1.3.39) (2025-03-02)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.45 ([#639](https://github.com/adobe/spacecat-shared/issues/639)) ([f1c2721](https://github.com/adobe/spacecat-shared/commit/f1c272173c9749095b6e4101bc8ccb7e436238dc))
+
+# [@adobe/spacecat-shared-content-client-v1.3.38](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.37...@adobe/spacecat-shared-content-client-v1.3.38) (2025-02-26)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.44 ([#629](https://github.com/adobe/spacecat-shared/issues/629)) ([de2daac](https://github.com/adobe/spacecat-shared/commit/de2daac12c176ccd30b499904c6688874c637281))
+
+# [@adobe/spacecat-shared-content-client-v1.3.37](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.36...@adobe/spacecat-shared-content-client-v1.3.37) (2025-02-25)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.43 ([#625](https://github.com/adobe/spacecat-shared/issues/625)) ([c3a4dbb](https://github.com/adobe/spacecat-shared/commit/c3a4dbbbe18047ee5a954cc7527d053db1d2edee))
+
+# [@adobe/spacecat-shared-content-client-v1.3.36](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.35...@adobe/spacecat-shared-content-client-v1.3.36) (2025-02-23)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.42 ([#623](https://github.com/adobe/spacecat-shared/issues/623)) ([4b20675](https://github.com/adobe/spacecat-shared/commit/4b20675adcae2a1405c85a9a36b567245eb95fd0))
+
+# [@adobe/spacecat-shared-content-client-v1.3.35](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.34...@adobe/spacecat-shared-content-client-v1.3.35) (2025-02-22)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#622](https://github.com/adobe/spacecat-shared/issues/622)) ([6552a61](https://github.com/adobe/spacecat-shared/commit/6552a61ebd8b83c0a1dec51d50d44cf3b97819c1))
+
+# [@adobe/spacecat-shared-content-client-v1.3.34](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.33...@adobe/spacecat-shared-content-client-v1.3.34) (2025-02-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.41 ([#621](https://github.com/adobe/spacecat-shared/issues/621)) ([78a7411](https://github.com/adobe/spacecat-shared/commit/78a7411958cf03376623735e3a85b110e1848eb0))
+
+# [@adobe/spacecat-shared-content-client-v1.3.33](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.32...@adobe/spacecat-shared-content-client-v1.3.33) (2025-02-20)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.3.40 ([#617](https://github.com/adobe/spacecat-shared/issues/617)) ([3f4ebe6](https://github.com/adobe/spacecat-shared/commit/3f4ebe66a8548d9a22cd72b0876266012bd15fc6))
+
 # [@adobe/spacecat-shared-content-client-v1.3.32](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.3.31...@adobe/spacecat-shared-content-client-v1.3.32) (2025-02-20)